+++ dashboard/app/api/index.ts
@@ -4218,6 +4218,108 @@ export const useReviewsControllerLeaveReview = <
   return useMutation(mutationOptions, queryClient);
 };
 
+export const companyControllerCreate = (
+  companyDetailsDto: CompanyDetailsDto,
+  signal?: AbortSignal,
+) => {
+  const formData = new FormData();
+  if (companyDetailsDto.name !== undefined) {
+    formData.append(`name`, companyDetailsDto.name);
+  }
+  if (companyDetailsDto.address !== undefined) {
+    formData.append(`address`, companyDetailsDto.address);
+  }
+  if (companyDetailsDto.contact_no !== undefined) {
+    formData.append(`contact_no`, companyDetailsDto.contact_no);
+  }
+  if (companyDetailsDto.website !== undefined) {
+    formData.append(`website`, companyDetailsDto.website);
+  }
+  if (companyDetailsDto.license_number !== undefined) {
+    formData.append(`license_number`, companyDetailsDto.license_number);
+  }
+  formData.append(`years_in_business`, companyDetailsDto.years_in_business);
+  formData.append(`employees_number`, companyDetailsDto.employees_number);
+  if (companyDetailsDto.created_at !== undefined) {
+    formData.append(`created_at`, companyDetailsDto.created_at);
+  }
+
+  return customFetcher<void>({
+    url: `/company`,
+    method: "POST",
+    headers: { "Content-Type": "multipart/form-data" },
+    data: formData,
+    signal,
+  });
+};
+
+export const getCompanyControllerCreateMutationOptions = <
+  TError = unknown,
+  TContext = unknown,
+>(options?: {
+  mutation?: UseMutationOptions<
+    Awaited<ReturnType<typeof companyControllerCreate>>,
+    TError,
+    { data: CompanyDetailsDto },
+    TContext
+  >;
+}): UseMutationOptions<
+  Awaited<ReturnType<typeof companyControllerCreate>>,
+  TError,
+  { data: CompanyDetailsDto },
+  TContext
+> => {
+  const mutationKey = ["companyControllerCreate"];
+  const { mutation: mutationOptions } = options
+    ? options.mutation &&
+      "mutationKey" in options.mutation &&
+      options.mutation.mutationKey
+      ? options
+      : { ...options, mutation: { ...options.mutation, mutationKey } }
+    : { mutation: { mutationKey } };
+
+  const mutationFn: MutationFunction<
+    Awaited<ReturnType<typeof companyControllerCreate>>,
+    { data: CompanyDetailsDto }
+  > = (props) => {
+    const { data } = props ?? {};
+
+    return companyControllerCreate(data);
+  };
+
+  return { mutationFn, ...mutationOptions };
+};
+
+export type CompanyControllerCreateMutationResult = NonNullable<
+  Awaited<ReturnType<typeof companyControllerCreate>>
+>;
+export type CompanyControllerCreateMutationBody = CompanyDetailsDto;
+export type CompanyControllerCreateMutationError = unknown;
+
+export const useCompanyControllerCreate = <
+  TError = unknown,
+  TContext = unknown,
+>(
+  options?: {
+    mutation?: UseMutationOptions<
+      Awaited<ReturnType<typeof companyControllerCreate>>,
+      TError,
+      { data: CompanyDetailsDto },
+      TContext
+    >;
+  },
+  queryClient?: QueryClient,
+): UseMutationResult<
+  Awaited<ReturnType<typeof companyControllerCreate>>,
+  TError,
+  { data: CompanyDetailsDto },
+  TContext
+> => {
+  const mutationOptions = getCompanyControllerCreateMutationOptions(options);
+
+  return useMutation(mutationOptions, queryClient);
+};
+
 export const companyControllerUpload = (
   id: string,
   uploadDocDto: UploadDocDto,
@@ -4300,83 +4402,6 @@ export const useCompanyControllerUpload = <
   TContext
 > => {
   const mutationOptions = getCompanyControllerUploadMutationOptions(options);
-
-  return useMutation(mutationOptions, queryClient);
-};
-
-export const companyControllerCreate = (
-  formData: FormData,
-  signal?: AbortSignal,
-) => {
-  return customFetcher<CommonResponseDto>({
-    url: `/company`,
-    method: "POST",
-    headers: { "Content-Type": "multipart/form-data" },
-    data: formData,
-    signal,
-  });
-};
-
-export const getCompanyControllerCreateMutationOptions = <
-  TError = unknown,
-  TContext = unknown,
->(options?: {
-  mutation?: UseMutationOptions<
-    Awaited<ReturnType<typeof companyControllerCreate>>,
-    TError,
-    { data: FormData },
-    TContext
-  >;
-}): UseMutationOptions<
-  Awaited<ReturnType<typeof companyControllerCreate>>,
-  TError,
-  { data: FormData },
-  TContext
-> => {
-  const mutationKey = ["companyControllerCreate"];
-  const { mutation: mutationOptions } = options
-    ? options.mutation &&
-      "mutationKey" in options.mutation &&
-      options.mutation.mutationKey
-      ? options
-      : { ...options, mutation: { ...options.mutation, mutationKey } }
-    : { mutation: { mutationKey } };
-
-  const mutationFn: MutationFunction<
-    Awaited<ReturnType<typeof companyControllerCreate>>,
-    { data: FormData }
-  > = (props) => {
-    const { data } = props ?? {};
-
-    return companyControllerCreate(data);
-  };
-
-  return { mutationFn, ...mutationOptions };
-};
-
-export type CompanyControllerCreateMutationResult = NonNullable<
-  Awaited<ReturnType<typeof companyControllerCreate>>
->;
-export type CompanyControllerCreateMutationBody = FormData;
-export type CompanyControllerCreateMutationError = unknown;
-
-export const useCompanyControllerCreate = <TError = unknown, TContext = unknown>(
-  options?: {
-    mutation?: UseMutationOptions<
-      Awaited<ReturnType<typeof companyControllerCreate>>,
-      TError,
-      { data: FormData },
-      TContext
-    >;
-  },
-  queryClient?: QueryClient,
-): UseMutationResult<
-  Awaited<ReturnType<typeof companyControllerCreate>>,
-  TError,
-  { data: FormData },
-  TContext
-> => {
-  const mutationOptions = getCompanyControllerCreateMutationOptions(options);
 
   return useMutation(mutationOptions, queryClient);
 };
